various fixes & additions
- fix sentinel setup in replication mode
- add custom STARTPORT option to delete mode i.e. ./redis-generator.sh delete 6579
- add debug output for delete mode

+++ redis-generator.sh
@@ -83,32 +83,90 @@ genredis_del() {
   for (( p=0; p <= $NUMBER; p++ ))
     do
       REDISPORT=$(($STARTPORT+$p))
+      SENTPORT=$(($REDISPORT+10000))
     if [ -f "/usr/lib/systemd/system/redis${REDISPORT}.service" ]; then
       echo "-------------------------------------------------------"
       echo "Deleting redis${REDISPORT}.service ..."
-      systemctl disable redis${REDISPORT}.service
-      systemctl stop redis${REDISPORT}.service
-      rm -rf "/usr/lib/systemd/system/redis${REDISPORT}.service"
+      if [[ "$DEBUG_REDISGEN" = [yY] ]]; then
+        echo "systemctl disable redis${REDISPORT}.service"
+        echo "systemctl stop redis${REDISPORT}.service"
+        echo "rm -rf "/usr/lib/systemd/system/redis${REDISPORT}.service""
+      else
+        systemctl disable redis${REDISPORT}.service
+        systemctl stop redis${REDISPORT}.service
+        rm -rf "/usr/lib/systemd/system/redis${REDISPORT}.service"
+      fi
     fi
     if [ -f "/etc/redis${REDISPORT}/redis${REDISPORT}.conf" ]; then
-      rm -rf "/etc/redis${REDISPORT}/redis${REDISPORT}.conf"
+      if [[ "$DEBUG_REDISGEN" = [yY] ]]; then
+        echo "rm -rf "/etc/redis${REDISPORT}/redis${REDISPORT}.conf""
+      else
+        rm -rf "/etc/redis${REDISPORT}/redis${REDISPORT}.conf"
+      fi
     fi
     if [ -f "/var/log/redis/redis${REDISPORT}.log" ]; then
-      rm -rf "/var/log/redis/redis${REDISPORT}.log"
+      if [[ "$DEBUG_REDISGEN" = [yY] ]]; then
+        echo "rm -rf "/var/log/redis/redis${REDISPORT}.log""
+      else
+        rm -rf "/var/log/redis/redis${REDISPORT}.log"
+      fi
     fi
     if [ -f "/var/lib/redis/dump${REDISPORT}.rdb" ]; then
-      rm -rf "/var/lib/redis/dump${REDISPORT}.rdb"
+      if [[ "$DEBUG_REDISGEN" = [yY] ]]; then
+        echo "rm -rf "/var/lib/redis/dump${REDISPORT}.rdb""
+      else
+        rm -rf "/var/lib/redis/dump${REDISPORT}.rdb"
+      fi
     fi
     if [ -f "/var/lib/redis/appendonly${REDISPORT}.aof" ]; then
-      rm -rf "/var/lib/redis/appendonly${REDISPORT}.aof"
+      if [[ "$DEBUG_REDISGEN" = [yY] ]]; then
+        echo "rm -rf "/var/lib/redis/appendonly${REDISPORT}.aof""
+      else
+        rm -rf "/var/lib/redis/appendonly${REDISPORT}.aof"
+      fi
     fi
     if [ -f "${SENTDIR}/sentinel-${REDISPORT}.conf" ]; then
-      rm -rf "${SENTDIR}/sentinel-${REDISPORT}.conf"
+      if [[ "$DEBUG_REDISGEN" = [yY] ]]; then
+        echo "rm -rf "${SENTDIR}/sentinel-${REDISPORT}.conf""
+      else
+        rm -rf "${SENTDIR}/sentinel-${REDISPORT}.conf"
+      fi
     fi
     if [ -d "/var/lib/redis/sentinel_${REDISPORT}" ]; then
-      rm -rf "/var/lib/redis/sentinel_${REDISPORT}"
+      if [[ "$DEBUG_REDISGEN" = [yY] ]]; then
+        echo "rm -rf "/var/lib/redis/sentinel_${REDISPORT}""
+      else
+        rm -rf "/var/lib/redis/sentinel_${REDISPORT}"
+      fi
     fi
-    rm -rf "/var/lib/redis${REDISPORT}"
+    if [ -f "/var/log/redis/sentinel-${REDISPORT}.log" ]; then
+      if [[ "$DEBUG_REDISGEN" = [yY] ]]; then
+        echo "rm -rf "/var/log/redis/sentinel-${REDISPORT}.log""
+      else
+        rm -rf "/var/log/redis/sentinel-${REDISPORT}.log"
+      fi
+    fi
+    if [ -f "/var/run/redis/redis-sentinel-${SENTPORT}.pid" ]; then
+      if [[ "$DEBUG_REDISGEN" = [yY] ]]; then
+        echo "kill -9 $(cat "/var/run/redis/redis-sentinel-${SENTPORT}.pid")"
+      else
+        kill -9 $(cat "/var/run/redis/redis-sentinel-${SENTPORT}.pid")
+      fi
+    fi
+    if [ -f "/etc/init.d/sentinel_${SENTPORT}" ]; then
+      if [[ "$DEBUG_REDISGEN" = [yY] ]]; then
+        echo "rm -rf "/etc/init.d/sentinel_${SENTPORT}""
+      else
+        rm -rf "/etc/init.d/sentinel_${SENTPORT}"
+      fi
+    fi
+    if [ -d "/var/lib/redis${REDISPORT}" ]; then
+      if [[ "$DEBUG_REDISGEN" = [yY] ]]; then
+        echo "rm -rf "/var/lib/redis${REDISPORT}""
+      else
+        rm -rf "/var/lib/redis${REDISPORT}"
+      fi
+    fi
   done
   echo "Deletion completed"
   exit
@@ -210,9 +268,9 @@ genredis() {
             fi
           fi
           # sentinel setup for redis replication
-          if [[ "$SENTINEL_SETUP" && "$CLUSTER" = 'replication' && "$REDISPORT" = "$STARTPORT" ]]; then
+          if [[ "$SENTINEL_SETUP" = [yY] && "$CLUSTER" = 'replication' && "$REDISPORT" = "$STARTPORT" ]]; then
             SPORT="$STARTPORT"
-            SENTPORT=$((SPORT+10000))
+            SENTPORT=$(($SPORT+10000))
             echo
             echo "-----------------"
             echo "creating ${SENTDIR}/sentinel-${SPORT}.conf ..."
@@ -357,9 +415,9 @@ esac
             fi
           fi
           # sentinel setup for redis replication
-          if [[ "$SENTINEL_SETUP" && "$CLUSTER" = 'replication' && "$REDISPORT" = "$STARTPORT" ]]; then
+          if [[ "$SENTINEL_SETUP" = [yY] && "$CLUSTER" = 'replication' && "$REDISPORT" = "$STARTPORT" ]]; then
             SPORT="$STARTPORT"
-            SENTPORT=$((SPORT+10000))
+            SENTPORT=$(($SPORT+10000))
             echo
             echo "-----------------"
             echo "creating ${SENTDIR}/sentinel-${SPORT}.conf ..."
@@ -560,6 +618,10 @@ case "$1" in
     ;;
   delete )
     NUM=$2
+    CUSTOM_STARTPORT=$3
+    if [[ ! -z "$CUSTOM_STARTPORT" ]]; then
+      STARTPORT=$CUSTOM_STARTPORT
+    fi
     genredis_del $NUM
     ;;
   * )
@@ -570,13 +632,14 @@ case "$1" in
 
     echo "* prep - standalone prep command installs redis-cluster-tool"
     echo "* prepupdate - standalone prep update command updates redis-cluster-tool"
-    echo "* multi X - number of standalone redis instances to create"
-    echo "* clusterprep X - number of cluster enabled config instances"
+    echo "* multi X - no. of standalone redis instances to create"
+    echo "* clusterprep X - no. of cluster enabled config instances"
     echo "* clustermake 6 - to enable cluster mode + create cluster"
     echo "* clustermake 9 - flag to enable cluster mode + create cluster"
     echo "* replication X - create redis replication"
     echo "* replication X 6579 - create replication with custom start port 6579"
-    echo "* delete X - number of redis instances to delete"
+    echo "* delete X - no. of redis instances to delete"
+    echo "* delete X 6579 - no. of redis instances to delete + custom start port 6579"
     echo
     echo "$0 prep"
     echo "$0 prepupdate"
@@ -587,6 +650,7 @@ case "$1" in
     echo "$0 replication X"
     echo "$0 replication X 6579"
     echo "$0 delete X"
+    echo "$0 delete X 6579"
     ;;
 esac
 exit
